requirements.txt
boto3==1.26.112
botocore==1.29.112
cffi==1.15.1
cryptography==40.0.1
jmespath==1.0.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
